+++ packages/template-retail-react-app/app/components/_app/index.jsx
@@ -10,6 +10,7 @@ import PropTypes from 'prop-types'
 import {useHistory, useLocation} from 'react-router-dom'
 import {getAssetUrl} from 'pwa-kit-react-sdk/ssr/universal/utils'
 import {getAppOrigin} from 'pwa-kit-react-sdk/utils/url'
+import StorefrontPreview from 'pwa-kit-react-sdk/storefront-preview'
 
 // Chakra
 import {Box, useDisclosure, useStyleConfig} from '@chakra-ui/react'
@@ -170,6 +171,9 @@ const App = (props) => {
             >
                 <CategoriesProvider treeRoot={allCategories} locale={targetLocale}>
                     <CurrencyProvider currency={currency}>
+                        {/* For pwa-kit v2, we need to use this getToken prop to make sure it can work properly
+                         with Storefront Preview feature*/}
+                        <StorefrontPreview getToken={() => window.localStorage.getItem('token')} />
                         <Seo>
                             <meta name="theme-color" content={THEME_COLOR} />
                             <meta name="apple-mobile-web-app-title" content={DEFAULT_SITE_TITLE} />